Fixed coin network info providers Changed coin net hash rate type from 'long' to 'double'

+++ Msv.AutoMiner/Msv.AutoMiner.CoinInfoService/External/Data/CoinNetworkStatistics.cs
@@ -3,7 +3,7 @@
     public class CoinNetworkStatistics
     {
         public double Difficulty { get; set; }
-        public long NetHashRate { get; set; }
+        public double NetHashRate { get; set; }
         public long Height { get; set; }
         public double? BlockTimeSeconds { get; set; }
         public double? BlockReward { get; set; }
