+++ Msv.AutoMiner/Msv.AutoMiner.Data/CoinNetworkInfo.cs
@@ -12,7 +12,7 @@ namespace Msv.AutoMiner.Data
 
         public double Difficulty { get; set; }
 
-        public long NetHashRate { get; set; }
+        public double NetHashRate { get; set; }
 
         public double BlockReward { get; set; }
