+++ Msv.AutoMiner/Msv.AutoMiner.FrontEnd/Models/Coins/CoinDisplayModel.cs
@@ -9,7 +9,7 @@ namespace Msv.AutoMiner.FrontEnd.Models.Coins
 
         public double Difficulty { get; set; }
 
-        public long NetHashRate { get; set; }
+        public double NetHashRate { get; set; }
 
         public double BlockReward { get; set; }
